requirements.txt
flask
flask-cors
google-generativeai 
python-dotenv
firebase-admin